+++ query_counter.py
